Simplifies example even further.
