+++ .gitignore
@@ -1,3 +1,3 @@
-/node_modules
-.next
 .DS_Store
+.next
+node_modules
