+++ app/layout.tsx
@@ -1,18 +1,18 @@
-import './globals.css'
+import "./globals.css";
 
 export const metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
-}
+  title: "Next Cache Demo",
+  description: "A simple demo of Next.js cache",
+};
 
 export default function RootLayout({
   children,
 }: {
-  children: React.ReactNode
+  children: React.ReactNode;
 }) {
   return (
     <html lang="en">
       <body>{children}</body>
     </html>
-  )
+  );
 }
