+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "next-cache-demo-temp",
+  "name": "next-cache-demo",
   "version": "0.1.0",
   "private": true,
   "scripts": {
